refactor:removed repeated lines
